Drown your issues in Kool Aid (Updating K Music Store app to use EF/E API surface)
This updates the K version of Music Store to use the updated Entity Framework API surface from the data repro. The changes made previously to the models and controllers in the full .NET version are also included here to bring both versions into line. The code is still non-functional as it was before.