Fix EF InvalidOperationException on Startup
The code that does an 'UPSERT' of seed data on app start was querying
existing instances and attaching as modified of seed instances on the
same context. Now that we track query results this will throw because
the context has two instances of the same entity (same key value).

Swapping to use a temporary context to query existing data.