Setting the full path to load the Config.json.
Specifying just the name of the file works fine on selfhost or IISExpress. When app is hosted in IIS, Config.json is being searched in the w3wp3.exe's folder path resulting in an exception.
So fixing the path of Config.json by retrieving the application base path from AppDomain.