Enabling CoreCLR variation of the test
Value of the KRE_DEFAULT_LIB of the 'k test' process was flowing into Helios resulting in an exception. As the test is always running on desktop clr, the value of this variable always points to the desktop KRE path resulting in the exception. This fix will fix the issue.