Fixing a test failure due to a recent change.
Adding some delay for the selfhost tests. With the recent change of merging the dbs selfhost based tests start even before the data is populated into the db & application start. Adding a slight delay to allow the setup finish.
And fixing a test bug that was not surfacing. Was checking with a fixed album id. This worked before as the id created was always 453.