Added instructions on how to source dnvm.sh on OsX
When I followed the instructions above I ran into the ```-bash: dnvm: command not found```.

This is solved by executing ```source dnvm.sh```, which I added to the description.

I also added some instructions on how to put this into your bash-start-script so that it will be ready with every new Terminal session/window

Rephrased and changed backtics