automate the autobahn test suite (#101)
However, we only run that automation if the Autobahn test suite runner
(wstest) is installed. After this is in and we can verify it on the CI
we'll look at making it mandatory for building the repo.

Also includes test applications and scripts for running Autobahn suites
against arbitrary ASP.NET 4.6, ASP.NET Core and HttpListener
applications.