Remove explicit static constructors
Explicit static cctors cause the C# compiler to not mark types as
beforefieldinit, which means the JIT will add checks to each static
method and instance constructors of the type to make sure that the
static constructor was previously called. This can be avoided by
removing explicit static constructors and initializing static fields
inline.