Support partial-graph builds
This merges pre-existing packages from previous builds of Universe with the list of packages to be produced in the current run.

Pre-existing packages are expected to be in .deps/build/ and are expected to have only one version of a package id in that folder.

This also requires any additional lineups to be in .deps/lineups/ so they can be filtered from the Internal.AspNetCore.Universe.Lineup file