Add property to allow the SDK to disable shared framework trimming (#1130)
The MicrosoftNETPlatformLibrary property instructs the .NET Core SDK to treat a particular package as the shared framework platform. This affects how the SDK will trim references and publish output, determines how the runtimeconfig files are generated, and may affect how optimizations are preformed by other tools. In some installations of .NET Core, the ASP.NET Core shared framework is not available. This change adds properties to let the SDK determine on which platforms ASP.NET Core is enabled.