Workaround issue in TC with artifact dependencies
TC does not allow optional artifact dependencies. It will fail to build
if it cannot find an artifact in a folder pattern. This is problematic
for patches when we may not always produce assets in every folder. This
works around this by adding a dummy file to all expected output folders.