Code dump of component tests
This is a dump of the existing component codegen tests from
the Blazor effort.

I'm also adding the minimal Component types as a new shim to get things
to compile. This is similar to how our shims work already for MVC.

The next step will be add port functionality until the tests pass.
\n\nCommit migrated from https://github.com/dotnet/aspnetcore-tooling/commit/49f49dd9f04f62202f4f7b73afa79353e3fc95de