Update Razor SDK and fix build
chmod+x on the build script was missing

I think something wierd was going on with the dependency resolution for
Roslyn. I updated to a later release and then we got some warnings about
one of our code fixes. I addressed these in the same way we deal with
them in our other code fix.