SDK updates for components tooling
This is a bug fix for the RazorComponentWithTargetPath xaml rule. We
need to have separate targets for components and views.

The other fix here is a change to the set of workarounds we're using for
components currently. We need this project capability for the correct
project host to be used when loading the project.
\n\nCommit migrated from https://github.com/dotnet/aspnetcore-tooling/commit/78163f7e4f38937831cea526d682c30aad7bf80c